speed up cli
