+++ cohort_creator/_cli.py
@@ -2,28 +2,15 @@
 
 from __future__ import annotations
 
-import argparse
 import logging
 import sys
 from pathlib import Path
 from typing import Sequence
 
-import pandas as pd
-from datalad import api
 from rich_argparse import RichHelpFormatter
 
-from cohort_creator._browse import browse
 from cohort_creator._parsers import global_parser
-from cohort_creator._utils import (
-    get_bids_filter,
-    get_list_datasets_to_install,
-    load_dataset_listing,
-    load_participant_listing,
-    validate_dataset_types,
-)
-from cohort_creator.data._update import update
 from cohort_creator.logger import cc_logger
-from cohort_creator.main import construct_cohort, get_data, install_datasets
 
 cc_log = cc_logger()
 
@@ -45,21 +32,6 @@ def set_verbosity(verbosity: int | list[int]) -> None:
         logging.getLogger("datalad.gitrepo").setLevel(logging.ERROR)
 
 
-def _get_participant_listing_from_args(args: argparse.Namespace) -> pd.DataFrame | None:
-    if args.participant_listing is None:
-        return None
-    participant_listing = Path(args.participant_listing[0]).resolve()
-    return load_participant_listing(participant_listing=participant_listing)
-
-
-def create_yoda(output_dir: Path) -> None:
-    if not output_dir.exists():
-        api.create(path=output_dir, cfg_proc="yoda", result_renderer="disabled")
-    if not (output_dir / ".datalad").exists():
-        cc_log.info(f"Creating yoda dataset for output in: {output_dir}")
-        api.create(path=output_dir, cfg_proc="yoda", force=True, result_renderer="disabled")
-
-
 def cli(argv: Sequence[str] = sys.argv) -> None:
     """Entry point."""
     parser = global_parser(formatter_class=RichHelpFormatter)
@@ -73,14 +45,21 @@ def cli(argv: Sequence[str] = sys.argv) -> None:
     set_verbosity(verbosity)
 
     if args.command in ["browse"]:
+        from cohort_creator._browse import browse
+
         debug = getattr(args, "debug", False)
         browse(debug=debug)
         exit(0)
 
     if args.command in ["update"]:
+        from cohort_creator.data._update import update
+
         debug = getattr(args, "debug", True)
         update(debug=debug)
         exit(0)
+
+    from cohort_creator._run import _get_participant_listing_from_args
+    from cohort_creator._utils import load_dataset_listing, validate_dataset_types
 
     output_dir = Path(args.output_dir[0]).resolve()
 
@@ -92,6 +71,8 @@ def cli(argv: Sequence[str] = sys.argv) -> None:
     dataset_listing = load_dataset_listing(dataset_listing=args.dataset_listing)
 
     if args.command in ["install", "all"]:
+        from cohort_creator._run import _execute_install, create_yoda
+
         create_yoda(output_dir)
         (output_dir / "sourcedata").mkdir(exist_ok=True, parents=True)
         _execute_install(dataset_listing, args, output_dir)
@@ -110,9 +91,13 @@ def cli(argv: Sequence[str] = sys.argv) -> None:
         # TODO handle case when several tasks are passed
         task = task[0]
 
+    from cohort_creator._run import _return_bids_filter
+
     bids_filter = _return_bids_filter(args=args)
 
     if args.command in ["get", "all"]:
+        from cohort_creator.main import get_data
+
         jobs = args.jobs
         if isinstance(jobs, list):
             jobs = jobs[0]
@@ -131,6 +116,8 @@ def cli(argv: Sequence[str] = sys.argv) -> None:
         exit(0)
 
     if args.command in ["copy", "all"]:
+        from cohort_creator.main import construct_cohort
+
         skip_group_mriqc = bool(args.skip_group_mriqc)
         construct_cohort(
             output_dir=output_dir,
@@ -146,30 +133,5 @@ def cli(argv: Sequence[str] = sys.argv) -> None:
         exit(0)
 
 
-def _execute_install(
-    dataset_listing: pd.DataFrame, args: argparse.Namespace, output_dir: Path
-) -> None:
-    participant_listing = _get_participant_listing_from_args(args)
-
-    datasets_to_install = get_list_datasets_to_install(
-        dataset_listing=dataset_listing, participant_listing=participant_listing
-    )
-
-    generate_participant_listing = getattr(args, "generate_participant_listing", False)
-
-    if participant_listing is None:
-        generate_participant_listing = True
-
-    install_datasets(
-        datasets=datasets_to_install,
-        output_dir=output_dir,
-        dataset_types=args.dataset_types,
-        generate_participant_listing=generate_participant_listing,
-    )
-
-
-def _return_bids_filter(args: argparse.Namespace) -> dict[str, dict[str, dict[str, str]]] | None:
-    if args.bids_filter_file is None:
-        return None
-    bids_filter_file = Path(args.bids_filter_file[0]).resolve()
-    return get_bids_filter(bids_filter_file=bids_filter_file) if bids_filter_file.exists() else None
+# if __name__ == "__main__":
+#     cli(['cohort_creator', '--help'])
